Add quince search to deal with horizon effect
