+++ src/ai/c_modules/hashkeys.c
