+++ src/ai/c_modules/init.c
