+++ src/ai/c_modules/misc.c
