+++ src/ai/c_modules/validate.c
